+++ src/php/filesystem/realpath.js
@@ -12,28 +12,28 @@ module.exports = function realpath (path) {
   }
 
   var p = 0
-  var arr = [] /* Save the root, if not given */
-  var r = this.window.location.href /* Avoid input failures */
+  var arr = [] // Save the root, if not given
+  var r = this.window.location.href // Avoid input failures
 
-  /* Check if there's a port in path (like 'http://') */
+  // Check if there's a port in path (like 'http://')
   path = (path + '').replace('\\', '/')
   if (path.indexOf('://') !== -1) {
     p = 1
   }
 
-  /* Ok, there's not a port in path, so let's take the root */
+  // Ok, there's not a port in path, so let's take the root
   if (!p) {
     path = r.substring(0, r.lastIndexOf('/') + 1) + path
   }
 
-  /* Explode the given path into it's parts */
-  arr = path.split('/') /* The path is an array now */
-  path = [] /* Foreach part make a check */
-  for (var k in arr) { /* This is'nt really interesting */
+  // Explode the given path into it's parts
+  arr = path.split('/') // The path is an array now
+  path = [] // Foreach part make a check
+  for (var k in arr) { // This is'nt really interesting
     if (arr[k] === '.') {
       continue
     }
-    /* This reduces the realpath */
+    // This reduces the realpath
     if (arr[k] === '..') {
       /* But only if there more than 3 parts in the path-array.
        * The first three parts are for the uri */
@@ -41,16 +41,16 @@ module.exports = function realpath (path) {
         path.pop()
       }
     } else {
-      /* This adds parts to the realpath
-       * But only if the part is not empty or the uri
-       * (the first three parts ar needed) was not
-       * saved */
+      // This adds parts to the realpath
+      // But only if the part is not empty or the uri
+      // (the first three parts ar needed) was not
+      // saved
       if ((path.length < 2) || (arr[k] !== '')) {
         path.push(arr[k])
       }
     }
   }
 
-  /* Returns the absloute path as a string */
+  // Returns the absloute path as a string
   return path.join('/')
 }